+++ src/components/report/ReportTable.tsx
@@ -30,7 +30,7 @@ import {
   capitalizeFirstLetter,
 } from "../../helpers";
 import { Group, Report, Source, Tag } from "../../objectTypes";
-import { editReport } from "../../api/reports";
+import { editReport } from "../../api/reports/index";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import TagsTypeahead from "../tag/TagsTypeahead";
 import { faArrowUpRightFromSquare } from "@fortawesome/free-solid-svg-icons";
