--- src/types/common.ts
@@ -1,3 +1,0 @@
-export interface hasId {
-  _id: string;
-}
